twit-to-mail.js - more css to prevent image truncation
Some email readers (well, Thunderbird) seem to embed the email HTML in
their own HTML headers, and this causes some odd behaviour I didn't
see in the context of a browser.  Specifically, the tweet appears to
be within a frame of fixed size, and the bottom of an image gets
clipped - but the clipping depends on the size of the mail body pane.
note, the size of the tweet and the pane don't match exactly!  So
there is still a scroll bar.

This fix works empirically in Thunderbird, and in Firefox / Chrome
given the HTML used by Thunderbird.  Untested in other mail clients /
browsers.